% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_mess.r
\name{create_mess}
\alias{create_mess}
\title{Creates message string for reporting during procedures}
\usage{
create_mess(
  mess,
  h = integer(0),
  indent = integer(0),
  prefix = "",
  h_marks = "-",
  h_prefix = character(),
  h_prefix_sep = " ",
  pretty = getOption("disambr_mess_pretty"),
  mess_color = "green",
  h_prefix_color = "blue",
  ...
)
}
\arguments{
\item{mess}{Message to report. If prefixed by h \code{h_marks} it will be ouline of level \code{h}}

\item{h}{Forse specific ouline level of message}

\item{indent}{Forse indentation}

\item{prefix}{Add overal prefix}

\item{h_marks}{Marks that sets outline. Default is "-". Can be many characters, e.g. "-*#".}

\item{h_prefix}{Character vector of prefixes for each outline level}

\item{h_prefix_sep}{Separator between \code{h_prefix} and \code{mess}}

\item{pretty}{Whether to use "crayon" package for pretty printing}

\item{mess_color}{Color of message}

\item{h_prefix_color}{Color of ouline prefix}

\item{...}{Here we can pass \code{verbose} argument from upper functions. Default is TRUE}
}
\value{
Message string
}
\description{
Creates message string for reporting during procedures
}
