% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{dhms}
\alias{dhms}
\title{Formats time difference as X days HH:MM:SS}
\usage{
dhms(t)
}
\arguments{
\item{t}{time diff}
}
\value{
formatted time diff string
}
\description{
from https://stackoverflow.com/questions/27312292
}
