% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\docType{package}
\name{disambr-package}
\alias{disambr}
\alias{disambr-package}
\title{disambr: NEDR - Named Entity Disambiguation in R}
\description{
NEDR - Named Entity Disambiguation in R
}
\details{
AND EVA Algorithm
This is work in progress. Please, file an issues or suggestion if you have any.
}
\author{
\strong{Maintainer}: Stanislav A. Vlasov \email{stanislav.a.vlasov@gmail.com}

Authors:
\itemize{
  \item Olmo R. van den Akker \email{ovdakker@gmail.com}
  \item Sacha Epskamp \email{sacha.epskamp@gmail.com}
}

}
\keyword{internal}
