% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\name{disambr.get.different.authors}
\alias{disambr.get.different.authors}
\title{Returns sets of people ids that are defenetely different based on co-authorship}
\usage{
disambr.get.different.authors(sets, procedures = NULL)
}
\arguments{
\item{sets}{}

\item{procedures}{}
}
\value{

}
\description{
Returns sets of people ids that are defenetely different based on co-authorship
}
