% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\name{disambr.get.similar.initials}
\alias{disambr.get.similar.initials}
\title{Returns set of people with similar initials}
\usage{
disambr.get.similar.initials(sets, procedures = NULL)
}
\arguments{
\item{sets}{}

\item{procedures}{}
}
\value{

}
\description{
Returns set of people with similar initials
}
