% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\name{disambr.get.similar.last.names}
\alias{disambr.get.similar.last.names}
\title{Returns set of people with similar last names}
\usage{
disambr.get.similar.last.names(sets, procedures = NULL)
}
\arguments{
\item{sets}{}

\item{procedures}{}
}
\value{

}
\description{
Returns set of people with similar last names
}
