% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_add_set_attr}
\alias{disambr_add_set_attr}
\title{Add disambr attribures to focal set from template set and update some of them}
\usage{
disambr_add_set_attr(
  focal_set,
  template_set = NULL,
  ...,
  attr_names_prefix = "disambr_set_"
)
}
\arguments{
\item{focal_set}{Set}

\item{template_set}{Set to inhirit attributes from}

\item{...}{
  Arguments passed on to \code{\link[=disambr_set_attr]{disambr_set_attr}}
  \describe{
    \item{\code{check_attr_names_prefix}}{Whether to check for short names of attributes. See \code{attr_names_prefix}.}
  }}

\item{attr_names_prefix}{"disambr_set_" by default. Only prefixed by it will be copied from \code{template_set}}
}
\value{

}
\description{
It adds attributes by reference (with setattr {data.table}), i.e. without making a copy
}
\details{
It updates:
\itemize{
\item time stamp (disambr_set_st) to current
\item disambr_set_file to NULL
\item adds to disambr_set_recipe the calling procedure
\item adds disambr_set_duration
}
}
