% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_filter_authors_by_researcher_ids}
\alias{disambr_filter_authors_by_researcher_ids}
\title{Subsets WoS authors table with Researcher IDs}
\usage{
disambr_filter_authors_by_researcher_ids(
  sets,
  file_path = system.file("testdata", "tekles-bornmann-researcher-ids.txt", package =
    "disambr")
)
}
\arguments{
\item{sets}{WoS data}

\item{file_path}{path to Researcher IDs. The default list of Researcher IDs is taken from Tekles & Bornmann (2019) for reproducing their test samples (see reference)}
}
\value{
updated WoS data
}
\description{
This procedure does not alter sets attributes. Just filters WoS author table.
}
\references{
Tekles, A., & Bornmann, L. (2019). Author name disambiguation of bibliometric data: A comparison of several unsupervised approaches. ArXiv:1904.12746. http://arxiv.org/abs/1904.12746
}
