% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_get_last_unstrong_set}
\alias{disambr_get_last_unstrong_set}
\title{Gets last set from sets with strength <= 0.5 and excludes from this set all sets with strength of 1}
\usage{
disambr_get_last_unstrong_set(sets)
}
\arguments{
\item{sets}{sets}
}
\value{
set or NULL
}
\description{
Gets last set from sets with strength <= 0.5 and excludes from this set all sets with strength of 1
}
