% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_get_output_set}
\alias{disambr_get_output_set}
\title{Gets output set from sets (in case we already made it)}
\usage{
disambr_get_output_set(
  sets,
  get_output_set = getOption("disambr_get_output_set"),
  attr_names_prefix = "disambr_set_"
)
}
\arguments{
\item{sets}{sets}

\item{get_output_set}{Whether to search for output set. Default is not.}

\item{attr_names_prefix}{prefix for attributes}
}
\value{
NULL or output set
}
\description{
Gets output set from sets (in case we already made it)
}
