% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_in_sets}
\alias{disambr_in_sets}
\title{Checks if sets with certain attribures are present}
\usage{
disambr_in_sets(
  sets,
  ...,
  match_attr_value_parcially = FALSE,
  check_attr_names_prefix = TRUE,
  attr_names_prefix = "disambr_set_"
)
}
\arguments{
\item{sets}{Sets to filter on sets attributes}

\item{...}{Named sets attributes to filter \code{sets} on}

\item{match_attr_value_parcially}{whether attribute values can be matched partially}

\item{check_attr_names_prefix}{Whether to check for short names of attributes. See \code{attr_names_prefix}.}

\item{attr_names_prefix}{If name does not start with this prefix (default is 'disambr_set_'), it will add this prefix before attribute name.}
}
\value{
logical vector of length \code{length(sets)}
}
\description{
Checks if sets with certain attribures are present
}
