% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_make_wos_tsv_author_year_citations}
\alias{disambr_make_wos_tsv_author_year_citations}
\title{Makes citations table by matching first author - year keys. This, of course, can produce false positives matched}
\usage{
disambr_make_wos_tsv_author_year_citations(pub_table, ref_table)
}
\arguments{
\item{pub_table}{WoS publications data table}

\item{ref_table}{WoS references data table}
}
\value{
citation table
}
\description{
Makes citations table by matching first author - year keys. This, of course, can produce false positives matched
}
