% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_make_wos_tsv_authors}
\alias{disambr_make_wos_tsv_authors}
\title{Parses all WoS field related to authors and makes author table}
\usage{
disambr_make_wos_tsv_authors(
  wos_data_table,
  list_of_author_fields = c("author_order", "author_short_name", "author_initials",
    "author_last_name", "author_full_name", "author_first_names",
    "author_first_full_name", "author_email", "author_researcher_id", "author_orcid",
    "author_affiliations"),
  ...
)
}
\arguments{
\item{wos_data_table}{WoS tsv export data.table}

\item{list_of_author_fields}{list of data to parse to construct (if not all selected as by default it will save time)}

\item{...}{set verbose if needed here}
}
\value{
author table
}
\description{
Relevant fields that are parsed (as in Web of Science Field Tags 2018-06-27):
\itemize{
\item AU
\item AF (full names)
\item C1 adresses
\item RP reprint address (one you contact for reprint copy)
\item EM emails
\item RI researcher ID
\item OI ORCID Identifier (Open Researcher and Contributor ID)
}
}
\details{
Fields that are not parsed
\itemize{
\item BA book - not parsed
\item BF book - not parsed
\item CA gp group author (usually organization or group name) - not parsed
\item BE editors - not parsed
}
}
