% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_make_wos_tsv_publications}
\alias{disambr_make_wos_tsv_publications}
\title{Make set of publications}
\usage{
disambr_make_wos_tsv_publications(tables_list, recipes)
}
\arguments{
\item{tables_list}{list of tables}

\item{recipes}{recipes}
}
\value{
set
}
\description{
Make set of publications
}
