% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_make_wos_tsv_references}
\alias{disambr_make_wos_tsv_references}
\title{Make references table from WoS tsv data}
\usage{
disambr_make_wos_tsv_references(wos_data_table, ...)
}
\arguments{
\item{wos_data_table}{WoS data}

\item{...}{set verbose here if neededj}
}
\value{
references table
}
\description{
Make references table from WoS tsv data
}
