% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_merge_authors_if_citing_others_papers}
\alias{disambr_merge_authors_if_citing_others_papers}
\title{Make a set of matched authors bases on the cases when one author cites the others paper.}
\usage{
disambr_merge_authors_if_citing_others_papers(
  sets,
  match_refrerences_by_name_year = TRUE
)
}
\arguments{
\item{sets}{Sets}

\item{match_refrerences_by_name_year}{Whether to check citations based on first author name and year pair in addition to machich citations based on DOI}
}
\description{
Make a set of matched authors bases on the cases when one author cites the others paper.
}
