% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_merge_authors_with_common_keywords}
\alias{disambr_merge_authors_with_common_keywords}
\title{Makes set of authors with number of keywords in their papers in common}
\usage{
disambr_merge_authors_with_common_keywords(sets, keywords_in_common = 3)
}
\arguments{
\item{sets}{Sets}

\item{keywords_in_common}{number of keywords in common}
}
\value{
Sets with new set appended
}
\description{
Makes set of authors with number of keywords in their papers in common
}
