% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_merge_authors_with_same_emails}
\alias{disambr_merge_authors_with_same_emails}
\title{Makes set of matched authors bases on same email addresses}
\usage{
disambr_merge_authors_with_same_emails(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{
Sets with new attached
}
\description{
Makes set of matched authors bases on same email addresses
}
