% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_merge_authors_with_same_researcher_ids}
\alias{disambr_merge_authors_with_same_researcher_ids}
\title{For testing. Makes so called ground truth set based on same research IDs}
\usage{
disambr_merge_authors_with_same_researcher_ids(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{
Sets with new set appended
}
\description{
For testing. Makes so called ground truth set based on same research IDs
}
