% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_mess.r
\name{disambr_mess}
\alias{disambr_mess}
\title{Report a message with message()}
\usage{
disambr_mess(mess, h_prefix = c("disambr:", "-"), ...)
}
\arguments{
\item{mess}{Message to report. If prefixed by h \code{h_marks} it will be ouline of level \code{h}}

\item{h_prefix}{Character vector of prefixes for each outline level}

\item{...}{
  Arguments passed on to \code{\link[=create_mess]{create_mess}}
  \describe{
    \item{\code{h}}{Forse specific ouline level of message}
    \item{\code{indent}}{Forse indentation}
    \item{\code{prefix}}{Add overal prefix}
    \item{\code{h_marks}}{Marks that sets outline. Default is "-". Can be many characters, e.g. "-*#".}
    \item{\code{h_prefix_sep}}{Separator between \code{h_prefix} and \code{mess}}
    \item{\code{pretty}}{Whether to use "crayon" package for pretty printing}
    \item{\code{mess_color}}{Color of message}
    \item{\code{h_prefix_color}}{Color of ouline prefix}
  }}
}
\value{
Same as \code{message} returns
}
\description{
Report a message with message()
}
