% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_messaging.r
\name{disambr_message_start}
\alias{disambr_message_start}
\title{Post a starting message for disambr procedure. Records time started in \code{disambr_start_time} variable in its \code{parent.frame()}}
\usage{
disambr_message_start(start_mess_prefix = "Making set -", ...)
}
\arguments{
\item{start_mess_prefix}{Prefix for staring message}

\item{...}{
  Arguments passed on to \code{\link[=disambr_message]{disambr_message}}
  \describe{
    \item{\code{mess}}{Message to report. If prefixed by h \code{h_marks} it will be ouline of level \code{h}}
    \item{\code{h_prefix}}{Character vector of prefixes for each outline level}
  }}
}
\value{
time started
}
\description{
Post a starting message for disambr procedure. Records time started in \code{disambr_start_time} variable in its \code{parent.frame()}
}
