% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_read.r
\name{disambr_read}
\alias{disambr_read}
\title{Reads the data for disambiguation}
\usage{
disambr_read(
  files_path,
  save_sets_as = NULL,
  save_sets_dir = "disambr-data",
  use_time_stamp = FALSE
)
}
\arguments{
\item{files_path}{Path to data. You can specify almost everything}
}
\value{

}
\description{
Reads the data for disambiguation
}
