% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_read.r
\name{disambr_read_file}
\alias{disambr_read_file}
\title{Reads file based on file extention}
\usage{
disambr_read_file(f)
}
\arguments{
\item{f}{full file path name}
}
\value{
data
}
\description{
Reads file based on file extention
}
