% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_read.r
\name{disambr_read_tsv_wos}
\alias{disambr_read_tsv_wos}
\title{Reads WoS tsv export file and makes disambr set out of it (just adding some attributes to the data.table)}
\usage{
disambr_read_tsv_wos(f, header)
}
\arguments{
\item{f}{path}

\item{header}{header}
}
\value{
disambr set
}
\description{
Reads WoS tsv export file and makes disambr set out of it (just adding some attributes to the data.table)
}
