% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_save_set}
\alias{disambr_save_set}
\title{Save set and adds file attribute}
\usage{
disambr_save_set(
  set_to_save,
  save_set_as = getOption("disambr_save_as"),
  save_set_prefix = getOption("disambr_save_set_prefix"),
  save_set_dir = getOption("disambr_save_set_dir"),
  use_time_stamp = getOption("disambr_save_set_time_stamp")
)
}
\arguments{
\item{set_to_save}{set}

\item{save_set_as}{if TRUE the the file name is made from disambr_set_name attribute}

\item{save_set_prefix}{file prefix, default is "disambr-set."}

\item{save_set_dir}{file dir, default is "disambr-sets-rds" in current directory}

\item{use_time_stamp}{add time stamps at the end of file name, adds by default}
}
\value{
file namej
}
\description{
Save set and adds file attribute
}
