% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_set_attr}
\alias{disambr_set_attr}
\title{Adds attribures to the set with data.table::setattr}
\usage{
disambr_set_attr(
  focal_set,
  ...,
  check_attr_names_prefix = TRUE,
  attr_names_prefix = "disambr_set_"
)
}
\arguments{
\item{focal_set}{Set to add attribute to}

\item{...}{Named attributes}

\item{check_attr_names_prefix}{Whether to check for short names of attributes. See \code{attr_names_prefix}.}

\item{attr_names_prefix}{If name does not start with this prefix (default is 'disambr_set_'), it will add this prefix before attribute name.}
}
\value{
\code{focal_set}
}
\description{
Adds attribures to the set with data.table::setattr
}
