% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_cite_others_paper}
\alias{disambr_set_cite_others_paper}
\title{Make a set of matched authors bases on the cases when one author cites the others paper.}
\usage{
disambr_set_cite_others_paper(sets, match_refrerences_by_name_year = TRUE)
}
\arguments{
\item{sets}{Sets}

\item{match_refrerences_by_name_year}{Whether to check citations based on first author name and year pair in addition to machich citations based on DOI}
}
\value{

}
\description{
Make a set of matched authors bases on the cases when one author cites the others paper.
}
