% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_cite_self_citation}
\alias{disambr_set_cite_self_citation}
\title{Make a set of matched authors based on cases when one author cites others self citation. Self-citations here are detected based on DOI.}
\usage{
disambr_set_cite_self_citation(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{
Sets with new appended
}
\description{
Make a set of matched authors based on cases when one author cites others self citation. Self-citations here are detected based on DOI.
}
