% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_common_references}
\alias{disambr_set_common_references}
\title{Make set of authors that have number of references in common}
\usage{
disambr_set_common_references(sets, references_in_common = 3)
}
\arguments{
\item{sets}{Sets}

\item{references_in_common}{number of references in common}
}
\value{
Sets with new set
}
\description{
Make set of authors that have number of references in common
}
