% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_on_same_paper}
\alias{disambr_set_on_same_paper}
\title{Makes sets of co-authors assuming that all authors on paper are different person.}
\usage{
disambr_set_on_same_paper(sets)
}
\arguments{
\item{sets}{Sests}
}
\value{
Updated sets
}
\description{
Makes sets of co-authors assuming that all authors on paper are different person.
}
