% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_same_affiliation}
\alias{disambr_set_same_affiliation}
\title{Makes set of matched authors based on same affiliation}
\usage{
disambr_set_same_affiliation(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{
Sets with new attached
}
\description{
Makes set of matched authors based on same affiliation
}
