% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_same_email}
\alias{disambr_set_same_email}
\title{Makes set of matched authors bases on same email addresses}
\usage{
disambr_set_same_email(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{
Sets with new attached
}
\description{
Makes set of matched authors bases on same email addresses
}
