% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_similar_initials}
\alias{disambr_set_similar_initials}
\title{Makes set of similar authors based on their initials}
\usage{
disambr_set_similar_initials(sets, maxDist = 1)
}
\arguments{
\item{sets}{Sets}

\item{maxDist}{see max_dist in \code{match_fuzzy}}
}
\value{
Sets with new appended
}
\description{
Makes set of similar authors based on their initials
}
