% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_set_tekles_bornmann}
\alias{disambr_set_tekles_bornmann}
\title{Subsets WoS authors table with Researcher IDs used in Tekles & Bornmann (2019). Meant for reproducability.}
\usage{
disambr_set_tekles_bornmann(
  sets,
  file_path = "../data/tekles-bornmann-researcher-ids.txt"
)
}
\arguments{
\item{sets}{WoS data}

\item{file_path}{path to Researcher IDs}
}
\value{
updated WoS data
}
\description{
This procedure does not alter sets attributes. Just filters WoS author table.
}
\details{
Reference: Tekles, A., & Bornmann, L. (2019). Author name disambiguation of bibliometric data: A comparison of several unsupervised approaches. ArXiv:1904.12746 \link{Cs}. http://arxiv.org/abs/1904.12746
}
