% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_stats.r
\name{disambr_stats}
\alias{disambr_stats}
\title{Make a statistics for a collection of sets (specific algorithm) and writes to the file in dir}
\usage{
disambr_stats(
  sets = NULL,
  sets_dir = getOption("disambr_save_set_dir"),
  name = sets_dir,
  remove_data = TRUE,
  save_rds = TRUE
)
}
\arguments{
\item{sets}{collection of sets}

\item{sets_dir}{dir with collection of sets if first arg is not provided, writes results here}

\item{name}{name}

\item{remove_data}{whether to remove first 4 sets (assumed to be a data) when sets arg is set}

\item{save_rds}{save_rds}
}
\value{

}
\description{
Make a statistics for a collection of sets (specific algorithm) and writes to the file in dir
}
