% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_stats.r
\name{disambr_stats_table}
\alias{disambr_stats_table}
\title{Makes comparative table with statistics}
\usage{
disambr_stats_table(sets_dir)
}
\arguments{
\item{sets_dir}{dir}
}
\value{

}
\description{
Makes comparative table with statistics
}
