% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_sets.r
\name{disambr_subsets}
\alias{disambr_subsets}
\title{Filters list of sets}
\usage{
disambr_subsets(
  sets_list,
  attribute_value_list,
  which_to_return = c("all", "first", "last"),
  negate_subsets = FALSE
)
}
\arguments{
\item{sets_list}{list of sets}

\item{attribute_value_list}{list of attribute values where list elements name correspond attribute names used for filtering sets}

\item{which_to_return}{whether to return "all", "first" or "last" set from filtered sets}

\item{negate_subsets}{whether to return sets that was not matched insted}
}
\value{
list of sets or set if \code{which_to_return} is ethier "first" or "last"
}
\description{
Filters list of sets
}
