% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_af}
\alias{disambr_wos_tsv_parse_af}
\title{Parses AF (author full name) column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_af(name)
}
\arguments{
\item{name}{(a record string from AF column)?}
}
\value{
Data.table
}
\description{
Parses AF (author full name) column of WoS saved records export
}
