% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_au}
\alias{disambr_wos_tsv_parse_au}
\title{Parses AU column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_au(author_name)
}
\arguments{
\item{record.au}{a record string from AU column}
}
\value{
data.table
}
\description{
Parses AU column of WoS saved records export
}
