% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_c1}
\alias{disambr_wos_tsv_parse_c1}
\title{Parses C1 (author adress/affiliation) column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_c1(record_c1, table_af = NULL)
}
\arguments{
\item{record_c1}{a record string from RP column}

\item{table_af}{Table}
}
\value{
Data.table with two columns -  author_name and affiliations
}
\description{
Parses C1 (author adress/affiliation) column of WoS saved records export
}
