% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_cr}
\alias{disambr_wos_tsv_parse_cr}
\title{Parses WoS CR (Cited References) record into separate talbe}
\usage{
disambr_wos_tsv_parse_cr(references)
}
\arguments{
\item{references}{CR (Cited References) record, i.e., just one row}
}
\value{
references table
}
\description{
Parses WoS CR (Cited References) record into separate talbe
}
