% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_em}
\alias{disambr_wos_tsv_parse_em}
\title{Parses EM (email) column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_em(emails, record_au_table, record_rp_table)
}
\arguments{
\item{record_au_table}{a data_tabe after parsing AU column with disambr_wos_tsv_parse_au}

\item{record_rp_table}{a data_tabe after parsing RP column with disambr_wos_tsv_parse_rp}

\item{record_em}{a record string from EM column}
}
\value{
Data.table with columns - author_name, affiliations and email
}
\description{
Parses EM (email) column of WoS saved records export
}
