% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_oi}
\alias{disambr_wos_tsv_parse_oi}
\title{Parses OI column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_oi(authors, table_af = NULL)
}
\arguments{
\item{authors}{a record string from OI column}

\item{table_af}{Table}
}
\value{
data.table
}
\description{
Parses OI column of WoS saved records export
}
