% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_rp}
\alias{disambr_wos_tsv_parse_rp}
\title{Parses RP (reprint author) column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_rp(record_rp)
}
\arguments{
\item{record_rp}{a record string from RP column}
}
\value{
Data.table with two columns -  author_name and affiliations
}
\description{
Parses RP (reprint author) column of WoS saved records export
}
