% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{get_file_extension}
\alias{get_file_extension}
\title{Extention extractor. Same as tools::file_ext but for NULL input returns NULL instead of logical(0).}
\usage{
get_file_extension(f)
}
\arguments{
\item{f}{file name}
}
\value{
extention
}
\description{
Extention extractor. Same as tools::file_ext but for NULL input returns NULL instead of logical(0).
}
