% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{get_upper_triangle_index}
\alias{get_upper_triangle_index}
\title{Returns the index of element in upper triangle of squared \code{n} by \code{n} matrix}
\usage{
get_upper_triangle_index(i, j, n)
}
\arguments{
\item{i}{row index}

\item{j}{column index}

\item{n}{the size of squared matrix}
}
\value{
An index as integer number. Index starts from 1
}
\description{
inspired from https://math.stackexchange.com/questions/2134011 but modified so
index starts from 1 as the original answer was for index starting from 0
}
