% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{match_fuzzy}
\alias{match_fuzzy}
\title{Fuzzy match all combinations of character vector}
\usage{
match_fuzzy(bank, method, max_dist, id_name)
}
\arguments{
\item{bank}{bank}

\item{method}{see method in stringdist}

\item{max_dist}{see maxDist in stringdist}

\item{id_name}{names that will be suffixed with _1 and _2}
}
\value{
data.table
}
\description{
Fuzzy match all combinations of character vector
}
