% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\name{parse.files.path}
\alias{parse.files.path}
\title{Returns vector of file paths from path(s) recursively}
\usage{
parse.files.path(files.path, recursive = TRUE)
}
\arguments{
\item{files.path}{Path(s) where the files are}

\item{recursive}{Whether to look in subfolders recursively}
}
\value{
Vector of file paths from path(s) recursively
}
\description{
Returns vector of file paths from path(s) recursively
}
