% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{parse_files_path}
\alias{parse_files_path}
\title{Returns vector of file paths from path(s) recursively}
\usage{
parse_files_path(files_path, recursive = TRUE)
}
\arguments{
\item{files_path}{Path(s) where the files are}

\item{recursive}{Whether to look in subfolders recursively}
}
\value{
Vector of file paths from path(s) recursively
}
\description{
Returns vector of file paths from path(s) recursively
}
