% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{read_to_utf8}
\alias{read_to_utf8}
\title{Reads file as UTF-8, convert it if other encoding is deteted}
\usage{
read_to_utf8(f, bytes_to_check = 2^14)
}
\arguments{
\item{f}{file path}

\item{bytes_to_check}{how long to check for encoding (save time for large files)}
}
\value{
file text as string
}
\description{
Reads file as UTF-8, convert it if other encoding is deteted
}
