% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{recode_return_characters}
\alias{recode_return_characters}
\title{Fixed end of line characters in wierd text}
\usage{
recode_return_characters(s, assoc.file = NA, verbose = FALSE)
}
\arguments{
\item{s}{text string}

\item{assoc.file}{file name where it came from}

\item{verbose}{Be chatty}
}
\value{
fixed sting
}
\description{
Fixed end of line characters in wierd text
}
