% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\name{stop.unless}
\alias{stop.unless}
\title{Stops process unless cond is true}
\usage{
stop.unless(
  cond,
  message.if.false = paste("cond in not TRUE"),
  stop.if.false = TRUE,
  return.if.true = TRUE,
  return.if.false = isFALSE(return.if.true)
)
}
\arguments{
\item{cond}{}

\item{message.if.false}{}

\item{stop.if.false}{}

\item{return.if.true}{}

\item{return.if.false}{}
}
\value{

}
\description{
Stops process unless cond is true
}
