% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_utils.r
\name{stop_unless}
\alias{stop_unless}
\title{Stops process unless cond is true}
\usage{
stop_unless(
  cond,
  message_if_false = paste("cond in not TRUE"),
  stop_if_false = TRUE,
  return_if_true = TRUE,
  return_if_false = isFALSE(return_if_true)
)
}
\arguments{
\item{cond}{condition to test}

\item{message_if_false}{message_if_false}

\item{stop_if_false}{stop_if_false}

\item{return_if_true}{return_if_true}

\item{return_if_false}{return_if_false}
}
\description{
Stops process unless cond is true
}
