% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr.r
\docType{package}
\name{disambr-package}
\alias{disambr}
\alias{disambr-package}
\title{disambr: disambr - Named Entity Disambiguation in R}
\description{
disambr - Named Entity Disambiguation in R
}
\details{
This package provides a framework for disambiguating named entities (e.g., authors in large bibliometric databases)

Package provides following main functions
\itemize{
\item disambr_read - reads WoS data
\item disambr_aev - implementation of the AEV algorithm (van den Akker et al., 2020) for Web of Science author disambiguation.
}

This is work in progress. Please, file an issues or suggestion if you have any.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stasvlasov/disambr}
  \item Report bugs at \url{https://github.com/stasvlasov/disambr/issues}
}

}
\author{
\strong{Maintainer}: Stanislav A. Vlasov \email{stanislav.a.vlasov@gmail.com}

Authors:
\itemize{
  \item Olmo R. van den Akker \email{ovdakker@gmail.com}
  \item Sacha Epskamp \email{sacha.epskamp@gmail.com}
}

}
\keyword{internal}
