% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_eva.r
\name{disambr_match_authors_if_sharing_coauthors}
\alias{disambr_match_authors_if_sharing_coauthors}
\title{Make a set of matched authors that share co-authors}
\usage{
disambr_match_authors_if_sharing_coauthors(sets)
}
\arguments{
\item{sets}{Sets}
}
\value{

}
\description{
Make a set of matched authors that share co-authors
}
