% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_stats.r
\name{disambr_stats}
\alias{disambr_stats}
\title{Make a statistics for a collection of sets (specific algorithm) and writes to the file in \code{sets_dir}}
\usage{
disambr_stats(
  sets = NULL,
  sets_dir = getOption("disambr_save_set_dir"),
  name = sets_dir,
  remove_data = TRUE,
  save_rds = TRUE
)
}
\arguments{
\item{sets}{Collection of sets}

\item{sets_dir}{If first arg is not provided search for collection of sets in this directory. The results are saved to this directory as well.}

\item{name}{Name of the collection of sets}

\item{remove_data}{whether to remove first 4 sets (assumed to be a data) when sets arg is set}

\item{save_rds}{save_rds}
}
\description{
Make a statistics for a collection of sets (specific algorithm) and writes to the file in \code{sets_dir}
}
