% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambr_wos_tsv.r
\name{disambr_wos_tsv_parse_ri}
\alias{disambr_wos_tsv_parse_ri}
\title{Parses RI column of WoS saved records export}
\usage{
disambr_wos_tsv_parse_ri(authors, table_af = NULL)
}
\arguments{
\item{authors}{a record string from RI column}

\item{table_af}{Table}
}
\value{
data.table
}
\description{
Parses RI column of WoS saved records export
}
